Automatically create `~/.clang-format` file
clang-format and thus clangd chooses to format C/C++ files with an
indent width of two spaces [1]. This config however assumes a global
default of four spaces for indentation. clang-format recursively
traverses to the parent directory looking for .clang-format or
_clang-format files. Thus we put a .clang-format file into the home
directory of the current user if the file does not exist already. We
recommend LLVM as the base style and four spaces for the indentation
width. These settings can be overridden on a per-project basis by
placing a .clang-format or _clang-format file in the project directory.

[1] https://clang.llvm.org/docs/ClangFormatStyleOptions.html